Replace daemon-reexec with daemon-reload
daemon-rexec is generally overkill with reload being preferred. 
snip from manpage:
This command is of little use except for debugging and package upgrades. Sometimes, it might be helpful as a heavy-weight daemon-reload. While the daemon is being reexecuted, all sockets systemd listening on behalf of user configuration will stay accessible.